Move helpers and selectors into same subfolder as Vulnerability Management integration tests
