+++ ui/apps/platform/cypress/integration/vulnmanagement/images.test.js
@@ -1,4 +1,3 @@
-import { selectors } from '../../constants/VulnManagementPage';
 import withAuth from '../../helpers/basicAuth';
 import { hasFeatureFlag } from '../../helpers/features';
 import {
@@ -6,6 +5,7 @@ import {
     callbackForPairOfAscendingNumberValuesFromElements,
     callbackForPairOfDescendingNumberValuesFromElements,
 } from '../../helpers/sort';
+
 import {
     getCountAndNounFromImageCVEsLinkResults,
     hasTableColumnHeadings,
@@ -13,7 +13,8 @@ import {
     verifyFixableCVEsLinkAndRiskAcceptanceTabs,
     verifySecondaryEntities,
     visitVulnerabilityManagementEntities,
-} from '../../helpers/vulnmanagement/entities';
+} from './vulnerabilityManagement.helpers';
+import { selectors } from './vulnerabilityManagement.selectors';
 
 const entitiesKey = 'images';
