+++ ui/apps/platform/cypress/integration/vulnmanagement/vulnerabilityManagement.helpers.js
@@ -1,9 +1,9 @@
-import { selectors } from '../../constants/VulnManagementPage';
-import { hasFeatureFlag } from '../features';
+import { hasFeatureFlag } from '../../helpers/features';
+import { visitFromLeftNavExpandable } from '../../helpers/nav';
+import { getRouteMatcherMapForGraphQL, interactAndWaitForResponses } from '../../helpers/request';
+import { visit } from '../../helpers/visit';
 
-import { visitFromLeftNavExpandable } from '../nav';
-import { getRouteMatcherMapForGraphQL, interactAndWaitForResponses } from '../request';
-import { visit } from '../visit';
+import { selectors } from './vulnerabilityManagement.selectors';
 
 let opnamesForDashboard = [
     'policiesCount',
