+++ ui/apps/platform/cypress/integration/vulnmanagement/vulnerabilityManagement.selectors.js
@@ -1,8 +1,8 @@
-import { selectors as tablePaginationSelectors } from './TablePagination';
-import panelSelectors from '../selectors/panel';
-import tableSelectors from '../selectors/table';
-import scopeSelectors from '../helpers/scopeSelectors';
-import navigationSelectors from '../selectors/navigation';
+import { selectors as tablePaginationSelectors } from '../../constants/TablePagination';
+import scopeSelectors from '../../helpers/scopeSelectors';
+import panelSelectors from '../../selectors/panel';
+import tableSelectors from '../../selectors/table';
+import navigationSelectors from '../../selectors/navigation';
 
 export const vmHomePageSelectors = {
     // TODO: remove this selector, after at least one sub-menu is added to Vuln Mgmt menu
